feat: fetch alternative metadata from DeArrow
After constructing the PlatformVideo, fetch data from DeArrow and populate the new `alternativeName` and `thumbnail.type` fields with the best title and thumbnail from the DeArrow data.

Also updates the TypeScript definitions and autocomplete script.

Apologies for the bad quality of the code. I just kind of went with the vibe of the rest of the codebase here.